some changes for compatibility with future plans
- move GetGodotClassTrampolines to GodotInternal
- Move trampoline collector types from
'Godot.Bridge.ScriptManagerBridge' to just 'Godot.Bridge'.

This is where these classes and methods are expected to be in a future
trim-safe support PR I'm preparing
