+++ modules/mono/glue/GodotSharp/GodotSharp/GodotSharp.csproj
@@ -56,6 +56,7 @@
   <ItemGroup>
     <Compile Include="Core\Aabb.cs" />
     <Compile Include="Core\Attributes\ExportToolButtonAttribute.cs" />
+    <Compile Include="Core\Bridge\Collectors.cs" />
     <Compile Include="Core\Bridge\GodotSerializationInfo.cs" />
     <Compile Include="Core\Bridge\MethodInfo.cs" />
     <Compile Include="Core\Bridge\GodotClassTrampolines.cs" />
